SCRUM-2 Añadido boton para unirse a la sala y highlight de sala seleccionada
