+++ src/styles/list.css
@@ -2,6 +2,10 @@
     width: 100%;
 }
 
+.table-fixed tbody tr.selected td {
+    background-color: gray;
+}
+
 .table-fixed thead,
 .table-fixed tbody {
     display: block;
